pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "Proprietary License. All Rights Reserved"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
sentry-sdk = {extras = ["fastapi"], version = "^1.38.0"}
gunicorn = "^21.2.0"
sqlmodel = "0.0.8"
python-dotenv = "^1.0.0"
itsdangerous = "^2.1.2"
email-validator = "1.3.1"
requests = "^2.31.0"
aiohttp = "^3.9.1"
babel = "2.12.1"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
psycopg2-binary = "^2.9.9"
starlette = "0.27.0"
jinja2 = "^3.1.3"
python-multipart = "^0.0.6"
pytz = "^2024.1"
sqlalchemy-utils = "^0.41.1"
pandas = "^2.2.1"
openpyxl = "^3.1.2"
pytest = "^8.3.5"
httpx = "^0.28.1"
pytest-asyncio = "^0.25.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = ["production: mark test as a production test"]
